♻️💥moved `AutoGrid` to `autogrid.py` and fixed tests
