+++ tests/test_editgrid.py
@@ -11,7 +11,7 @@ from ipyautoui.custom.buttonbars import CrudButtonBar
 from ipyautoui.demo_schemas import EditableGrid
 from ipyautoui.autoipywidget import AutoObject
 from ipyautoui.automapschema import _init_model_schema
-from ipyautoui.custom.editgrid import GridSchema, AutoObjectFiltered
+from ipyautoui.custom.editgrid import AutoObjectFiltered
 from pydantic import BaseModel, Field
 import typing as ty
 
@@ -54,223 +54,6 @@ class TestButtonBar:
             backward=backward,
         )
         assert button_bar.fn_add() == "ADD"
-
-
-class TestAutoGridInitData:
-    def test_empty_grid(self):
-        class Cols(BaseModel):
-            string: str = Field(aui_column_width=100)
-            floater: float = Field(aui_column_width=70, aui_sig_fig=3)
-
-        class DataFrameSchema(BaseModel):
-            """no default"""
-
-            __root__: ty.List[Cols] = Field(format="dataframe")
-
-        # initiate empty grid
-        grid = AutoGrid(schema=DataFrameSchema)
-        assert grid._data["data"] == []
-        assert grid._data["schema"]["fields"] == [
-            {"name": "index", "type": "integer"},  # NOTE: unable to detect type
-            {"name": "String", "type": "string"},
-            {"name": "Floater", "type": "string"},  # NOTE: unable to detect type
-            {"name": "ipydguuid", "type": "integer"},
-        ]
-
-    def test_assign_default_at_root(self):
-        # get default data from top-level schema defaults
-
-        class Cols(BaseModel):
-            string: str = Field("string", aui_column_width=100)
-            floater: float = Field(3.14, aui_column_width=70, aui_sig_fig=3)
-
-        class DataFrameSchema(BaseModel):
-            """default."""
-
-            __root__: ty.List[Cols] = Field(
-                [Cols(string="test", floater=1.5).dict()], format="dataframe"
-            )
-
-        grid = AutoGrid(schema=DataFrameSchema)
-        assert grid._data["data"] == [
-            {"index": 0, "String": "test", "Floater": 1.5, "ipydguuid": 0}
-        ]
-
-    def test_pass_data_as_kwarg(self):
-        # get default data passed as kwarg, titles as column headers
-        class Cols(BaseModel):
-            string: str = Field("string", aui_column_width=100)
-            floater: float = Field(3.14, aui_column_width=70, aui_sig_fig=3)
-
-        class DataFrameSchema(BaseModel):
-            """no default. but properties have default"""
-
-            __root__: ty.List[Cols] = Field(format="dataframe")
-
-        df = pd.DataFrame([{"String": "test2", "Floater": 2.2}])
-        grid3 = AutoGrid(schema=DataFrameSchema, data=df)
-        assert grid3._data["data"] == [
-            {"index": 0, "String": "test2", "Floater": 2.2, "ipydguuid": 0}
-        ]
-
-    def test_pass_data_as_kwarg_map_titles(self):
-        # get default data passed as kwarg, keys as column headers. maps to titles
-
-        class Cols(BaseModel):
-            string: str = Field("string", aui_column_width=100)
-            floater: float = Field(3.14, aui_column_width=70, aui_sig_fig=3)
-
-        class DataFrameSchema(BaseModel):
-            """no default. but properties have default"""
-
-            __root__: ty.List[Cols] = Field(format="dataframe")
-
-        df = pd.DataFrame([{"string": "test2", "floater": 2.2}])
-        grid4 = AutoGrid(schema=DataFrameSchema, data=df)
-        assert grid4._data["data"] == [
-            {"index": 0, "String": "test2", "Floater": 2.2, "ipydguuid": 0}
-        ]
-        print("done")
-
-    def test_reset_multiindex_data_with_init_data(self):
-        # get default data passed as kwarg, keys as column headers. maps to titles
-
-        class TestProperties(BaseModel):
-            string: str = Field(column_width=100, section="a")
-            floater: float = Field(1.5, column_width=70, aui_sig_fig=3, section="b")
-            inty: int = Field(1, section="b")
-
-        class TestGridSchema(BaseModel):
-            """no default"""
-
-            __root__: ty.List[TestProperties] = Field(
-                [TestProperties(string="string").dict()],
-                format="dataframe",
-                datagrid_index_name=("section", "title"),
-            )
-
-        df = pd.DataFrame([{"string": "test2", "floater": 2.2, "inty": 1}])
-        gr = AutoGrid(schema=TestGridSchema, data=df)
-
-        assert gr._data["data"] == [
-            {
-                ("index", ""): 0,
-                ("a", "String"): "test2",
-                ("b", "Floater"): 2.2,
-                ("b", "Inty"): 1,
-                ("ipydguuid", ""): 0,
-            }
-        ]
-
-        df1 = pd.DataFrame([{"string": "test2", "floater": 2.2, "inty": 1}] * 2)
-        gr.data = gr._init_data(df1)
-
-        assert gr._data["data"] == [
-            {
-                ("a", "String"): "test2",
-                ("b", "Floater"): 2.2,
-                ("b", "Inty"): 1,
-                ("ipydguuid", ""): 0,
-                ("index", ""): 0,
-            },
-            {
-                ("a", "String"): "test2",
-                ("b", "Floater"): 2.2,
-                ("b", "Inty"): 1,
-                ("ipydguuid", ""): 1,
-                ("index", ""): 1,
-            },
-        ]
-
-        gr.transposed = True
-        df = gr._init_data(
-            pd.DataFrame([{"string": "test2", "floater": 2.2, "inty": 1}])
-        )
-        print("done")
-
-    @pytest.mark.parametrize("transposed", [True, False])
-    def test_order_index(self, transposed: bool):
-        class Cols(BaseModel):
-            string: str = Field(aui_column_width=100)
-            floater: float = Field(aui_column_width=70, aui_sig_fig=3)
-
-        class DataFrameSchema(BaseModel):
-            """no default"""
-
-            __root__: ty.List[Cols] = Field(format="dataframe")
-
-        order = (
-            "floater",
-            "string",
-        )
-        # Test without data passed
-        grid_without_data = AutoGrid(
-            schema=DataFrameSchema, transposed=transposed, order=order
-        )
-        # Test with data passed
-        grid_with_data = AutoGrid(
-            schema=DataFrameSchema,
-            data=pd.DataFrame([Cols(string="test", floater=2.5).dict()]),
-            transposed=transposed,
-            order=order,
-        )
-        if transposed:
-            assert tuple(grid_without_data.data.index) == tuple(
-                [grid_without_data.map_name_index.get(name) for name in order]
-            )
-            assert tuple(grid_with_data.data.index) == tuple(
-                [grid_with_data.map_name_index.get(name) for name in order]
-            )
-        else:
-            assert tuple(grid_without_data.data.columns) == tuple(
-                [grid_without_data.map_name_index.get(name) for name in order]
-            )
-            assert tuple(grid_with_data.data.columns) == tuple(
-                [grid_with_data.map_name_index.get(name) for name in order]
-            )
-
-    @pytest.mark.parametrize("transposed", [True, False])
-    def test_order_multi_index(self, transposed: bool):
-        class Cols(BaseModel):
-            string: str = Field(aui_column_width=100, title="String", section="a")
-            floater: float = Field(
-                aui_column_width=70, aui_sig_fig=3, title="Floater", section="a"
-            )
-
-        class DataFrameSchema(BaseModel):
-            """no default"""
-
-            __root__: ty.List[Cols] = Field(
-                format="dataframe",
-                datagrid_index_name=("section", "title"),
-            )
-
-        order = (
-            "floater",
-            "string",
-        )
-        # Test without data passed
-        grid_without_data = AutoGrid(schema=DataFrameSchema, order=order)
-        # Test with data passed
-        grid_with_data = AutoGrid(
-            schema=DataFrameSchema,
-            data=pd.DataFrame([Cols(string="test", floater=2.5).dict()]),
-            transposed=transposed,
-            order=order,
-        )
-        if transposed is True:
-            assert tuple(grid_without_data.data.index) == ()
-            assert tuple(grid_with_data.data.index) == tuple(
-                [grid_with_data.map_name_index.get(name) for name in order]
-            )
-        else:
-            assert tuple(grid_without_data.data.columns) == tuple(
-                [grid_without_data.map_name_index.get(name) for name in order]
-            )
-            assert tuple(grid_with_data.data.columns) == tuple(
-                [grid_with_data.map_name_index.get(name) for name in order]
-            )
-
 
 class TestEditGrid:
     def test_editgrid_change_data(self):
